Align deployment module dependencies
