+++ web/features/deployments/detail/index.tsx
@@ -10,7 +10,7 @@ import Link from '@/next/link'
 import { useSelectedLayoutSegment } from '@/next/navigation'
 import { CreateReleaseControl } from '../create-release'
 import { deploymentRouteAppInstanceIdAtom } from '../route-state'
-import { DeveloperApiHeaderSwitch } from './api-tokens/api-token-management/section'
+import { DeveloperApiHeaderSwitch } from './api-tokens/developer-api-header-switch'
 import { NewDeploymentHeaderAction } from './instances/header-actions/new-deployment-button'
 import { releasesLocalAtoms } from './releases/state'
 import { INSTANCE_DETAIL_TAB_KEYS, isInstanceDetailTabKey } from './tabs'
