+++ web/features/deployments/detail/overview/access-summary/access-status-section.tsx
@@ -9,7 +9,6 @@ import { SkeletonRectangle } from '@/app/components/base/skeleton'
 import Link from '@/next/link'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { DeploymentStatusBadge } from '../../../shared/ui/deployment-status-badge'
-import { OVERVIEW_CARD_CLASS_NAME, OVERVIEW_ICON_CLASS_NAME, OVERVIEW_INTERACTIVE_CARD_CLASS_NAME } from '../card-styles'
 
 type AccessStatusSectionProps = {
   accessChannels?: AccessChannels
@@ -31,6 +30,12 @@ type AccessStatusItem = {
 }
 
 const ACCESS_STATUS_SKELETON_KEYS = ['webapp', 'cli']
+const OVERVIEW_CARD_CLASS_NAME = 'rounded-xl border border-components-panel-border bg-components-panel-bg p-4'
+const OVERVIEW_INTERACTIVE_CARD_CLASS_NAME = cn(
+  OVERVIEW_CARD_CLASS_NAME,
+  'transition-colors hover:border-components-panel-border-subtle hover:bg-components-panel-on-panel-item-bg-hover focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-components-button-primary-bg',
+)
+const OVERVIEW_ICON_CLASS_NAME = 'flex size-8 shrink-0 items-center justify-center rounded-lg bg-background-section-burn text-text-tertiary'
 
 export function AccessStatusSection({ accessChannels }: AccessStatusSectionProps) {
   const { t } = useTranslation('deployments')
