+++ web/features/deployments/detail/overview/environment-status/environment-strip.tsx
@@ -11,10 +11,10 @@ import { openDeployDrawerAtom } from '../../../deploy-drawer/state'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { DeploymentEmptyState } from '../../../shared/components/empty-state'
 import { hasRuntimeInstanceDeployment } from '../../../shared/domain/runtime-status'
-import { OVERVIEW_CARD_CLASS_NAME } from '../card-styles'
 import { EnvironmentTile } from './environment-tile'
 
 const OVERVIEW_RUNTIME_INSTANCE_LIMIT = 4
+const OVERVIEW_CARD_CLASS_NAME = 'rounded-xl border border-components-panel-border bg-components-panel-bg p-4'
 
 type EnvironmentStripProps = {
   rows: EnvironmentDeployment[]
